# docs/openapi.yaml
openapi: 3.1.1
info:
  title: AgentFeedback API
  version: '1.0'
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: >-
    HTTP API of AgentFeedback, a service that stores write-once feedback from AI
    coding agents. The contract is this document together with the JSON Schemas
    under `schemas/` and the conformance kit under `conformance/`. This is the API
    of the next major release. Schemas use the custom keyword `x-max-bytes` for
    UTF-8 byte limits, since JSON Schema `maxLength` counts code points. The
    server serves this document at `/api/v1/openapi.json` in bundled form, with
    the referenced schema files inlined.
servers:
  - url: https://api.agentfeedback.io
    description: Hosted
  - url: http://127.0.0.1:8090
    description: Local
security:
  - bearer: []
  - apiKey: []
...
paths:
  /api/v1/submissions:
    post:
      tags: [submissions]
      operationId: createSubmission
      summary: Create a submission
      description: >-
        Accepts any JSON object. The envelope schema describes what is stored
        after inference and is never used to reject a body; deviations are
        normalised and reported as `warnings`. Returns 201 with `Location` for a
        new record, or 200 with `Location` of the existing record for a keyed
        replay of identical content or a keyless duplicate. A body that is not
        JSON or not an object is 400 `bad_request`. A reused `key` with
        different content is 409 `replay_mismatch`. A body over
        `limits.body_bytes` is 413 `request_too_large`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/envelope.v1.json'
      responses:
        '201':
          description: Created.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateResponse'}
              example:
                submission:
                  id: 44
                  uid: 019250f2-8c3e-7a4b-9d1e-2f3a4b5c6d7e
                  kind: friction
                  schema_version: 1
                  key: friction-workstation-a-019250f2
                  summary: README install step 3 references a flag that no longer exists
                  machine: workstation-a
                  model: claude-fable-5-1
                  harness: claude-code
                  project: example
                  occurred_at: '2026-09-27T09:58:12.000000Z'
                  context: {git_commit: a1b2c3d, client: agentfeedback/4.0.0}
                  payload: {category: documentation, details: The --fast flag was removed in 3.0., fix_status: applied, fix_ref: example@a1b2c3d}
                  content_hash: 9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a
                  created_at: '2026-09-27T10:00:01.123456Z'
                warnings: []
        '200':
          description: Replay of an existing record (keyed replay or keyless duplicate).
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    get:
      tags: [submissions]
      operationId: listSubmissions
      summary: List submissions
      description: >-
        All parameters are optional. Unknown parameter names, empty values and
        repeated singleton parameters are 400 `validation_error` naming the
        parameter. Rows are the record minus `payload` (present with
        `include=payload`). `total` counts every row matching the filters,
        ignoring cursors and `limit`, in the same read transaction. When
        `has_more` is true exactly one of `next_before_id` / `next_after_id` is
        set, matching the direction used.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - $ref: '#/components/parameters/SchemaVersion'
        - $ref: '#/components/parameters/Key'
        - $ref: '#/components/parameters/Machine'
        - $ref: '#/components/parameters/Model'
        - $ref: '#/components/parameters/Project'
        - $ref: '#/components/parameters/Harness'
        - $ref: '#/components/parameters/Category'
        - $ref: '#/components/parameters/FixStatus'
        - $ref: '#/components/parameters/ExcludeKind'
        - $ref: '#/components/parameters/Verdict'
        - $ref: '#/components/parameters/Processed'
        - $ref: '#/components/parameters/Redacted'
        - $ref: '#/components/parameters/ContentHash'
        - $ref: '#/components/parameters/Since'
        - $ref: '#/components/parameters/Until'
        - $ref: '#/components/parameters/On'
        - $ref: '#/components/parameters/Q'
        - name: before_id
          in: query
          description: Newest-first page of rows with `id < before_id`. Mutually exclusive with `after_id`.
          schema: {type: integer, minimum: 1}
        - name: after_id
          in: query
          description: Oldest-first page of rows with `id > after_id`. Mutually exclusive with `before_id`.
          schema: {type: integer, minimum: 0}
        - name: limit
          in: query
          description: 1–500, or 1–100 with `include=payload`; default 50. Out of range is 400.
          schema: {type: integer, minimum: 1, maximum: 500, default: 50}
        - name: include
          in: query
          description: '`payload`: rows carry `payload`.'
          schema: {type: string, enum: [payload]}
      responses:
        '200':
          description: A page of submissions.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListResponse'}
              example:
                submissions:
                  - id: 1188
                    uid: 019250f2-8c3e-7a4b-9d1e-2f3a4b5c6d7e
                    kind: friction
                    schema_version: 1
                    summary: README install step 3 references a flag that no longer exists
                    machine: workstation-a
                    project: example
                    content_hash: 9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a
                    created_at: '2026-09-27T10:00:01.123456Z'
                limit: 50
                total: 245
                has_more: true
                next_before_id: 1187
                next_after_id: null
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <key>`; scheme name case-insensitive. Either header authorises; neither shadows the other.'
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: '`X-Api-Key: <key>`. Either header authorises; neither shadows the other.'
  headers:
    XRequestId:
      description: Request identifier; present on every response.
      schema: {type: string}
    CacheControl:
      description: '`no-store` on every response of an authenticated route, success or error.'
      schema: {type: string, const: no-store}
    WWWAuthenticate:
      description: '`Bearer` on 401.'
      schema: {type: string}
    Location:
      description: Path of the stored record, e.g. `/api/v1/submissions/44`.
      schema: {type: string}
    Allow:
      description: Methods the route accepts.
      schema: {type: string}
...
    RetryAfter:
      description: Seconds to wait before retrying, when known.
      schema: {type: integer, minimum: 0}
  parameters:
...
    Kind:
      name: kind
      in: query
      description: Token, normalised like the field; `unknown` selects un-kinded rows.
      schema: {type: string}
    SchemaVersion:
      name: schema_version
      in: query
      description: Exact schema version.
      schema: {type: integer, minimum: 1}
    Key:
      name: key
      in: query
      description: Exact key.
      schema: {type: string}
    Machine:
      name: machine
      in: query
      description: Exact machine.
      schema: {type: string}
    Model:
      name: model
      in: query
      description: Exact model.
      schema: {type: string}
    Project:
      name: project
      in: query
      description: Exact project.
      schema: {type: string}
    Harness:
      name: harness
      in: query
      description: Harness token.
      schema: {type: string}
    Category:
      name: category
      in: query
      description: Friction payload `category`; other kinds never match.
      schema: {type: string}
    FixStatus:
      name: fix_status
      in: query
      description: Friction payload `fix_status`; other kinds never match.
      schema: {type: string}
    ExcludeKind:
      name: exclude_kind
      in: query
      description: Repeatable; rows of these kinds are omitted.
      style: form
      explode: true
      schema:
        type: array
        items: {type: string}
    Verdict:
      name: verdict
      in: query
      description: Verdict token.
      schema: {type: string}
    Processed:
      name: processed
      in: query
      description: '`false` is the queue.'
      schema: {type: boolean}
    Redacted:
      name: redacted
      in: query
      description: 'Default: both.'
      schema: {type: boolean}
    ContentHash:
      name: content_hash
      in: query
      description: Exact content hash.
      schema: {type: string, pattern: '^[0-9a-f]{64}$'}
    Since:
      name: since
      in: query
      description: RFC 3339, inclusive; `+` must be percent-encoded.
      schema: {type: string, format: date-time}
    Until:
      name: until
      in: query
      description: RFC 3339, inclusive; `+` must be percent-encoded.
      schema: {type: string, format: date-time}
    On:
      name: 'on'
      in: query
      description: Which timestamp `since`/`until` (and `bucket`) apply to.
      schema: {type: string, enum: [created_at, occurred_at], default: created_at}
    Q:
      name: q
      in: query
      description: >-
        Case-insensitive substring over `summary` and every string value in
        `payload` at any depth (keys, numbers and booleans are not searched;
        case folds by the simple Unicode lower-case mapping); unranked; at
        most 200 bytes.
      schema: {type: string}
  responses:
...
    BadRequest:
      description: >-
        `bad_request` (body not JSON, or not an object) or `validation_error`
        (reads, processing, import, path parameters).
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example:
            error: validation_error
            message: limit must be between 1 and 500
            request_id: 3f9c2a7e5b1d4c08
            details:
              - code: out_of_range
                pointer: '?limit'
                message: limit must be between 1 and 500
    Unauthorized:
      description: '`unauthorized`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        WWW-Authenticate: {$ref: '#/components/headers/WWWAuthenticate'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    MethodNotAllowed:
      description: '`method_not_allowed`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Allow: {$ref: '#/components/headers/Allow'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: '`replay_mismatch`: the key was already used with different content.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example:
            error: replay_mismatch
            message: key friction-workstation-a-019250f2 was already used with different content
            request_id: 3f9c2a7e5b1d4c09
            details:
              - code: key_reused
                pointer: /key
                message: key friction-workstation-a-019250f2 is stored as submission 44 with different content
                existing_id: 44
    TooLarge:
      description: '`request_too_large`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    RateLimited:
      description: '`rate_limited` (hosted service only).'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    InternalError:
      description: '`internal_error`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unavailable:
      description: '`unavailable`; `Retry-After` when known.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Timestamp:
      type: string
      format: date-time
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{6}Z$'
      description: RFC 3339 UTC with six fractional digits.
...
    Submission:
      type: object
      additionalProperties: false
      description: >-
        The stored record: the envelope after inference plus server fields.
        Optional members are omitted when unset, never sent as `null`. A
        redacted record (tombstone) has `payload: {"redacted": true}`, no
        `summary`, no `context`, and `redacted_at` set.
      required: [id, uid, kind, schema_version, payload, content_hash, created_at]
      properties:
        id: {type: integer, minimum: 1}
        uid: {type: string, format: uuid, description: UUIDv7 generated on create; preserved on import.}
        kind: {$ref: '../schemas/envelope.v1.json#/properties/kind'}
        schema_version: {$ref: '../schemas/envelope.v1.json#/properties/schema_version'}
        key: {$ref: '../schemas/envelope.v1.json#/properties/key'}
        summary: {$ref: '../schemas/envelope.v1.json#/properties/summary'}
        machine: {$ref: '../schemas/envelope.v1.json#/properties/machine'}
        model: {$ref: '../schemas/envelope.v1.json#/properties/model'}
        harness: {$ref: '../schemas/envelope.v1.json#/properties/harness'}
        project: {$ref: '../schemas/envelope.v1.json#/properties/project'}
        occurred_at: {$ref: '#/components/schemas/Timestamp'}
        context: {$ref: '../schemas/envelope.v1.json#/properties/context'}
        payload:
          $ref: '../schemas/envelope.v1.json#/properties/payload'
          description: >-
            Kind-specific content; see FrictionPayload
            (`schemas/kinds/friction.v1.json`) and ReviewPayload
            (`schemas/kinds/review.v1.json`). `{"redacted": true}` on a tombstone.
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        processed_at: {$ref: '#/components/schemas/Timestamp'}
        redacted_at: {$ref: '#/components/schemas/Timestamp'}
        verdict: {type: string}
        resolution: {type: string}
        ref: {type: string}
        processed_by: {type: string}
    SubmissionListItem:
      type: object
      additionalProperties: false
      description: The record minus `payload`; `payload` is present only with `include=payload`. Optional members are omitted, never `null`.
      required: [id, uid, kind, schema_version, content_hash, created_at]
      properties:
        id: {type: integer, minimum: 1}
        uid: {type: string, format: uuid}
        kind: {$ref: '../schemas/envelope.v1.json#/properties/kind'}
        schema_version: {$ref: '../schemas/envelope.v1.json#/properties/schema_version'}
        key: {$ref: '../schemas/envelope.v1.json#/properties/key'}
        summary: {$ref: '../schemas/envelope.v1.json#/properties/summary'}
        machine: {$ref: '../schemas/envelope.v1.json#/properties/machine'}
        model: {$ref: '../schemas/envelope.v1.json#/properties/model'}
        harness: {$ref: '../schemas/envelope.v1.json#/properties/harness'}
        project: {$ref: '../schemas/envelope.v1.json#/properties/project'}
        occurred_at: {$ref: '#/components/schemas/Timestamp'}
        context: {$ref: '../schemas/envelope.v1.json#/properties/context'}
        payload: {$ref: '../schemas/envelope.v1.json#/properties/payload'}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        processed_at: {$ref: '#/components/schemas/Timestamp'}
        redacted_at: {$ref: '#/components/schemas/Timestamp'}
        verdict: {type: string}
        resolution: {type: string}
        ref: {type: string}
        processed_by: {type: string}
    Warning:
      type: object
      additionalProperties: false
      required: [code, pointer, message]
      properties:
        code: {type: string}
        pointer: {type: string, description: 'RFC 6901 pointer into the stored record; query parameters use `?name`.'}
        message: {type: string}
        existing_id: {type: integer, minimum: 1, description: Only on the 409 `key_reused` detail.}
    Error:
      type: object
      required: [error, message, request_id]
      properties:
        error:
          type: string
          enum: [bad_request, validation_error, unauthorized, not_found, method_not_allowed, replay_mismatch, request_too_large, rate_limited, internal_error, unavailable]
        message: {type: string, description: 'Names the field or parameter, the value received when short, and the accepted range or shape.'}
        request_id: {type: string}
        details:
          type: array
          items: {$ref: '#/components/schemas/Warning'}
    CreateResponse:
      type: object
      required: [submission, warnings]
      properties:
        submission: {$ref: '#/components/schemas/Submission'}
        warnings:
          type: array
          items: {$ref: '#/components/schemas/Warning'}
    ListResponse:
      type: object
      required: [submissions, limit, total, has_more, next_before_id, next_after_id]
      properties:
        submissions:
          type: array
          items: {$ref: '#/components/schemas/SubmissionListItem'}
        limit: {type: integer, minimum: 1, maximum: 500}
        total: {type: integer, minimum: 0}
        has_more: {type: boolean}
        next_before_id: {type: [integer, 'null']}
        next_after_id: {type: [integer, 'null']}
...